Fix number of tosses larger than INT_MAX

+++ pp_hw2/part1/parallel_pi.cpp
@@ -14,7 +14,7 @@ void *thread_pi(void *data) {
     uniform_real_distribution<float> ran(0.0, 1.0);
 
     thread_data *d = (thread_data*) data;
-    int cnt = 0;
+    long long cnt = 0;
     for(int i = 0; i < d->tosses; i++) {
         double x = ran(engine), y = ran(engine);
         double distance = x * x + y * y;
@@ -32,23 +32,23 @@ int main(int argc, char* argv[]) {
     }
     long long num_threads, num_tosses, tosses_each;
     int rem;
-    num_threads = atoi(argv[1]);
-    num_tosses = atoi(argv[2]);
+    num_threads = atoll(argv[1]);
+    num_tosses = atoll(argv[2]);
     tosses_each = num_tosses / num_threads;
     rem = num_tosses % num_threads;
 
     pthread_t *ids = (pthread_t*) malloc(sizeof(pthread_t) * num_threads);
     thread_data *datas = (thread_data*) malloc(sizeof(thread_data) * num_threads);
 
-    for(int i = 0; i < num_threads; i++) {
+    for(long long i = 0; i < num_threads; i++) {
         datas[i].tosses = tosses_each + (i < rem);
         pthread_create(&ids[i], NULL, thread_pi, (void*) &datas[i]);
     }
     for(int i = 0; i < num_threads; i++)
         pthread_join(ids[i], NULL);
 
-    int res = 0;
-    for(int i = 0; i < num_threads; i++)
+    long long res = 0;
+    for(long long i = 0; i < num_threads; i++)
         res += datas[i].result;
     
     float pi = 4 * (double) res / (double) num_tosses;
